+++ app/src/main/java/com/orderMate/fragment/orderDetail/OrderDetailFragment.kt
@@ -1178,7 +1178,7 @@ class OrderDetailFragment : Fragment(), IOrderItemClickListener, ILineItemUpdate
                                     if (lineItem?.item?.id == itemId) {
                                         android.util.Log.d("ItemNoteReceivedDebug", "Setting note on lineItem: ${lineItem.id}")
                                         lineItem.note = note
-                                        lineItem.unitQty = quantity.toLong()
+                                        lineItem.unitQty = quantity
                                     }
                                 }
